requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
transformers==4.30.2
torch==2.0.1
numpy==1.25.2
scikit-learn==1.3.0
python-multipart==0.0.6
Pillow==10.0.0
